Refine workspace setup checklist UI
Split the mixed host implementation into focused components and hooks. Use semantic success styling for completion and status icons for pointer rows while preserving their accessible labels.

Refs ENG-1656
