+++ libs/client/onboarding/src/components/setup-checklist-completion.tsx
@@ -1,0 +1,159 @@
+import {Button} from '@shipfox/react-ui/button';
+import {Icon} from '@shipfox/react-ui/icon';
+import {Text} from '@shipfox/react-ui/typography';
+import {useEffect, useRef} from 'react';
+
+const JSDOM_USER_AGENT_RE = /jsdom/u;
+
+export function SetupChecklistCompletion({
+  showBurst,
+  onBurstComplete,
+  onDone,
+}: {
+  showBurst: boolean;
+  onBurstComplete?: (() => void) | undefined;
+  onDone?: (() => void) | undefined;
+}) {
+  return (
+    <div className="relative overflow-hidden border-b border-tag-success-border bg-tag-success-bg p-panel">
+      <ConfettiBurst active={showBurst} onComplete={onBurstComplete} />
+      <div className="relative flex items-center gap-group">
+        <Icon
+          name="checkCircleSolid"
+          className="size-20 shrink-0 text-tag-success-icon"
+          aria-hidden="true"
+        />
+        <div className="min-w-0 flex-1">
+          <Text size="sm" bold className="text-tag-success-text">
+            You're set up
+          </Text>
+          <Text size="xs" className="text-tag-success-text">
+            Your workspace is ready for its first workflow.
+          </Text>
+        </div>
+        <Button type="button" size="sm" variant="primary" onClick={onDone}>
+          Done
+        </Button>
+      </div>
+    </div>
+  );
+}
+
+function ConfettiBurst({
+  active,
+  onComplete,
+}: {
+  active: boolean;
+  onComplete?: (() => void) | undefined;
+}) {
+  const canvasRef = useRef<HTMLCanvasElement>(null);
+
+  useEffect(() => {
+    if (!active || typeof window === 'undefined') return;
+    let completed = false;
+    const finish = () => {
+      if (completed) return;
+      completed = true;
+      onComplete?.();
+    };
+
+    if (window.matchMedia?.('(prefers-reduced-motion: reduce)').matches) {
+      finish();
+      return;
+    }
+    if (typeof navigator !== 'undefined' && JSDOM_USER_AGENT_RE.test(navigator.userAgent)) {
+      finish();
+      return;
+    }
+
+    const canvas = canvasRef.current;
+    if (!canvas) {
+      finish();
+      return;
+    }
+    const bounds = canvas.getBoundingClientRect();
+    if (bounds.width === 0 && bounds.height === 0) {
+      finish();
+      return;
+    }
+    const context = canvas.getContext('2d');
+    if (!context) {
+      finish();
+      return;
+    }
+
+    const pixelRatio = Math.min(window.devicePixelRatio || 1, 2);
+    const width = Math.max(bounds.width, 1);
+    const height = Math.max(bounds.height, 1);
+    canvas.width = Math.floor(width * pixelRatio);
+    canvas.height = Math.floor(height * pixelRatio);
+    context.scale(pixelRatio, pixelRatio);
+
+    const styles = getComputedStyle(canvas);
+    const colors = [
+      styles.getPropertyValue('--color-background-accent-blue-base').trim(),
+      styles.getPropertyValue('--color-background-accent-purple-base').trim(),
+      styles.getPropertyValue('--color-background-accent-success-base').trim(),
+      styles.getPropertyValue('--color-background-accent-warning-base').trim(),
+    ].filter(Boolean);
+    if (colors.length === 0) {
+      finish();
+      return;
+    }
+    const palette = colors;
+    const particles = Array.from({length: 32}, (_, index) => ({
+      x: width / 2 + (Math.random() - 0.5) * width * 0.3,
+      y: height * 0.15,
+      vx: (Math.random() - 0.5) * 3,
+      vy: -(Math.random() * 3 + 2),
+      rotation: Math.random() * Math.PI,
+      size: Math.random() * 4 + 3,
+      color: palette[index % palette.length] ?? palette[0] ?? '',
+    }));
+    let frame = 0;
+    let startedAt = performance.now();
+
+    const draw = (now: number) => {
+      const elapsed = now - startedAt;
+      context.clearRect(0, 0, width, height);
+      context.globalAlpha = Math.max(0, 1 - elapsed / 1500);
+      for (const particle of particles) {
+        particle.vy += 0.12;
+        particle.vx *= 0.99;
+        particle.x += particle.vx;
+        particle.y += particle.vy;
+        particle.rotation += 0.12;
+        context.save();
+        context.translate(particle.x, particle.y);
+        context.rotate(particle.rotation);
+        context.fillStyle = particle.color;
+        context.fillRect(
+          -particle.size / 2,
+          -particle.size / 2,
+          particle.size,
+          particle.size * 0.6,
+        );
+        context.restore();
+      }
+      context.globalAlpha = 1;
+      if (elapsed < 1500) {
+        frame = requestAnimationFrame(draw);
+      } else {
+        finish();
+      }
+    };
+
+    startedAt = performance.now();
+    frame = requestAnimationFrame(draw);
+    return () => {
+      cancelAnimationFrame(frame);
+      context.clearRect(0, 0, width, height);
+    };
+  }, [active, onComplete]);
+
+  return (
+    <div aria-hidden="true" className="pointer-events-none absolute inset-0 size-full">
+      <canvas ref={canvasRef} className="size-full" />
+    </div>
+  );
+}
