+++ libs/client/onboarding/src/core/setup-checklist.test.ts
@@ -285,8 +285,7 @@ describe('deriveSetupChecklist', () => {
     const checklist = deriveSetupChecklist(input());
 
     expect(checklist.items.find((item) => item.id === 'tools')).toMatchObject({
-      purpose:
-        'Connect issue tracking, messaging, observability, or any Shipfox integration so workflows can react to events and agents can use them',
+      purpose: 'Connect issue tracking, messaging, observability, or any Shipfox integration',
       action: {href: '/settings/integrations'},
     });
   });
